Move languages to their own directory
Languages have some internals that are nice to keep next to the language
implementation itself, but we don't want to auto-export them from the
root mod.ts. By first moving all languages into their own directory we
can split out these internals from ending up in the public API.
